+++ backend/src/main/java/com/postpaf/Models/Post.java
@@ -9,10 +9,10 @@ public class Post {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "id_user", nullable = false)
-    private Integer idUser;
+    private Long idUser;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_user", insertable = false, updatable = false)
@@ -31,7 +31,7 @@ public class Post {
     public Post() {
     }
 
-    public Post(Integer idUser, String title, String contenu) {
+    public Post(Long idUser, String title, String contenu) {
         this.idUser = idUser;
         this.title = title;
         this.contenu = contenu;
@@ -39,19 +39,19 @@ public class Post {
     }
 
     // Getters et Setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getIdUser() {
+    public Long getIdUser() {
         return idUser;
     }
 
-    public void setIdUser(Integer idUser) {
+    public void setIdUser(Long idUser) {
         this.idUser = idUser;
     }
